# api/openapi.json
{
  "components": {
    "schemas": {
      "AdminCreateInviteRequest": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "default": 1,
            "title": "Count",
            "type": "integer"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "max_uses": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uses"
          }
        },
        "required": [
          "expires_at"
        ],
        "title": "AdminCreateInviteRequest",
        "type": "object"
      },
      "AdminCreateInvitesResponse": {
        "properties": {
          "invite_codes": {
            "items": {
              "type": "string"
            },
            "title": "Invite Codes",
            "type": "array"
          }
        },
        "required": [
          "invite_codes"
        ],
        "title": "AdminCreateInvitesResponse",
        "type": "object"
      },
      "AdminCreateProjectRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "study_settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Settings"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "AdminCreateProjectRequest",
        "type": "object"
      },
...
      "AdminParticipantItem": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "last_push_failure_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Push Failure At"
          },
          "last_push_success_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Push Success At"
          },
          "push_subscription_count": {
            "title": "Push Subscription Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "status",
          "created_at",
          "push_subscription_count"
        ],
        "title": "AdminParticipantItem",
        "type": "object"
      },
      "AdminParticipantsResponse": {
        "properties": {
          "participants": {
            "items": {
              "$ref": "#/components/schemas/AdminParticipantItem"
            },
            "title": "Participants",
            "type": "array"
          }
        },
        "required": [
          "participants"
        ],
        "title": "AdminParticipantsResponse",
        "type": "object"
      },
      "AdminProjectItem": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "member_count": {
            "title": "Member Count",
            "type": "integer"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "created_at",
          "member_count"
        ],
        "title": "AdminProjectItem",
        "type": "object"
      },
      "AdminProjectUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "active",
                  "paused",
                  "ended"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "AdminProjectUpdateRequest",
        "type": "object"
      },
      "AdminProjectsResponse": {
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/AdminProjectItem"
            },
            "title": "Projects",
            "type": "array"
          }
        },
        "required": [
          "projects"
        ],
        "title": "AdminProjectsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "DeviceJWT": {
        "description": "Device-signed ES256 JWT minted by the client. The JWT header must include kid set to the device id, and the aud claim must be h4ckath0n:http.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Hackathon app powered by h4ckath0n",
    "title": "h4ckath0n",
    "version": "0.1.9"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/projects": {
      "get": {
        "operationId": "admin_list_projects_admin_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProjectsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "DeviceJWT": []
          }
        ],
        "summary": "Admin List Projects",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "admin_create_project_admin_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Admin Create Project Admin Projects Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DeviceJWT": []
          }
        ],
        "summary": "Admin Create Project",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/projects/{project_id}": {
      "patch": {
        "operationId": "admin_update_project_admin_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminProjectUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Admin Update Project Admin Projects  Project Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DeviceJWT": []
          }
        ],
        "summary": "Admin Update Project",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/projects/{project_id}/export": {
      "get": {
        "operationId": "admin_project_export_admin_projects__project_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Project Export Admin Projects  Project Id  Export Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DeviceJWT": []
          }
        ],
        "summary": "Admin Project Export",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/projects/{project_id}/invites": {
      "post": {
        "operationId": "admin_create_invites_admin_projects__project_id__invites_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCreateInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCreateInvitesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DeviceJWT": []
          }
        ],
        "summary": "Admin Create Invites",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/projects/{project_id}/participants": {
      "get": {
        "operationId": "admin_project_participants_admin_projects__project_id__participants_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminParticipantsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DeviceJWT": []
          }
        ],
        "summary": "Admin Project Participants",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}